chore(#142): delete legacy core/utils/streaming.py
Removes the legacy chunked-IO streaming base (DataStreamer/HailDataStreamer/
StreamProcessor), the last remnant of the pre-#121 Streamer concept. The
supported concepts are the core/streamers Artifact ABCs and the self-contained
Annotator base (PR #150).

- gene_disease_table.GeneDiseaseTableStreamer: drop HailDataStreamer
  inheritance; self-contained __init__/setup/teardown mirroring the Annotator
  base (idempotent Hail init, no-op teardown, local DEFAULT_CHUNK_SIZE).
  ClinGen/GenCC/COSMIC subclasses unchanged.
- annotation_pipeline: relocate the StreamProcessor orchestrator (sole consumer
  ConfigurableAnnotationPipeline) into the module; permissive duck-typed hints.
- Retire the legacy ClinVar training-set pipeline (training_set.py +
  training_set_streamed.py) — no live callers, superseded by the #144
  TrainingSetBuilder.
- Tests: relocate the unrelated load_sample_chd_gene_set test to
  test_gene_sets_utils.py; swap the Mock(spec=...) base streamer in
  test_flexible_annotation.py.

Completes the deletion half of #142 (rename half merged in #150).
